Refactor and rename classes and repositories: update Msg to Notification, TbImpEtl to EtlTable, and related repository and service classes for improved clarity and consistency.

+++ src/main/java/com/wgzhao/addax/admin/dto/EtlBatchReq.java
@@ -12,5 +12,5 @@ import java.util.List;
 public class EtlBatchReq {
     private List<String> tids;
     private String flag;
-    private Long retryCnt;
+    private int retryCnt;
 }
